feat: support login

+++ server/api/v1/option/option_delete.go
@@ -14,6 +14,7 @@ import (
 
 type DeleteOptionRequest struct {
 	OptionID string `path:"id"`
+	Name     string `json:"name"`
 }
 
 func DeleteOption(ctx context.Context, c *app.RequestContext, req *DeleteOptionRequest) (bool, *types.PageInfo, error) {
@@ -23,6 +24,9 @@ func DeleteOption(ctx context.Context, c *app.RequestContext, req *DeleteOptionR
 	}
 	if option.Name == model.OptionNameSemesterStartDate || option.Name == model.OptionNameSemesterEndDate {
 		return false, nil, errors.New("cannot delete semester start or end date")
+	}
+	if string(option.Name) != req.Name {
+		return false, nil, errors.New("option name does not match")
 	}
 
 	_, err = query.OptionColl.Raw().DeleteOne(ctx, bson.M{"_id": option.ID})
